auto: changing return type

+++ exec/kpmeans.cpp
@@ -46,7 +46,7 @@ int main(int argc, char* argv[]) {
 
     std::string dist_type = "eucl";
     std::string centersfn = "";
-	unsigned max_iters=std::numeric_limits<unsigned>::max();
+	size_t max_iters=std::numeric_limits<size_t>::max();
 	std::string init = "kmeanspp";
 	unsigned nthread = kpmbase::get_num_omp_threads();
 	int num_opts = 0;
@@ -145,7 +145,7 @@ int main(int argc, char* argv[]) {
         printf("Read data!\n");
 
         unsigned* p_clust_asgns = new unsigned [nrow];
-        unsigned* p_clust_asgn_cnt = new unsigned [k];
+        size_t* p_clust_asgn_cnt = new size_t [k];
 
         if (NULL == p_centers) // We have no preallocated centers
             p_centers = new double [k*ncol];
